refactorization ready with new folder structure
